add check for entity exist or not

+++ service/book/book_service.go
@@ -6,6 +6,8 @@ import (
 	"github.com/krogertechnology/krogo/pkg/krogo"
 	"github.com/nitesh-zs/bookshelf-api/model"
 	"github.com/nitesh-zs/bookshelf-api/store"
+
+	"net/http"
 )
 
 type svc struct {
@@ -34,6 +36,20 @@ func (s svc) Create(ctx *krogo.Context, book *model.Book, user *model.User) (*mo
 		return nil, errors.InvalidParam{Param: []string{"invalid body request"}}
 	}
 
+	// check if book already exist
+	exists, err := s.store.IsExist(ctx, nil, &book.RegNum)
+	if err != nil {
+		return nil, errors.DB{Err: err}
+	}
+
+	if exists {
+		return nil, errors.MultipleErrors{
+			StatusCode: http.StatusConflict,
+			Errors:     []error{errors.EntityAlreadyExists{}},
+		}
+	}
+
+	// generating new book
 	resp, err := s.store.Create(ctx, book)
 
 	if err != nil {
@@ -46,6 +62,15 @@ func (s svc) Create(ctx *krogo.Context, book *model.Book, user *model.User) (*mo
 func (s svc) Update(ctx *krogo.Context, book *model.Book, user *model.User) (*model.BookRes, error) {
 	if book == nil {
 		return nil, errors.InvalidParam{Param: []string{"invalid body request"}}
+	}
+
+	exists, err := s.store.IsExist(ctx, &book.ID, nil)
+	if err != nil {
+		return nil, errors.DB{Err: err}
+	}
+
+	if !exists {
+		return nil, errors.EntityNotFound{Entity: "id"}
 	}
 
 	resp, err := s.store.Update(ctx, book)
@@ -62,10 +87,13 @@ func (s svc) Delete(ctx *krogo.Context, id uuid.UUID, user *model.User) error {
 		return errors.InvalidParam{Param: []string{"id"}}
 	}
 
-	_, err := s.store.GetByID(ctx, id)
-
+	exists, err := s.store.IsExist(ctx, &id, nil)
 	if err != nil {
-		return err
+		return errors.DB{Err: err}
+	}
+
+	if !exists {
+		return errors.EntityNotFound{Entity: "id"}
 	}
 
 	err = s.store.Delete(ctx, id)
